Move the build_maxtext_lock_cli.py to the root level to enable faster iteration.
